+++ gwcelery/tasks/gcn/__init__.py
@@ -6,12 +6,13 @@ References
 
 .. [GCN] https://gcn.gsfc.nasa.gov
 """
+import contextlib
 import socket
 import struct
+import time
 
-from celery import Task
 from celery.utils.log import get_task_logger
-from celery_eternal import EternalProcessTask
+from celery_eternal import EternalTask, EternalProcessTask
 from gcn import get_notice_type, NoticeType
 import gcn
 
@@ -20,57 +21,61 @@ from ..core import DispatchHandler
 
 log = get_task_logger(__name__)
 
-_size_struct = struct.Struct("!I")
 
+@app.task(base=EternalTask, bind=True)
+def broker(self):
+    """Single-client VOEvent broker for sending notices to GCN.
 
-class _SendTask(Task):
-
-    def __init__(self):
-        self.conn = None
-
-
-@app.task(queue='voevent', base=_SendTask, ignore_result=True, bind=True,
-          autoretry_for=(socket.error,), default_retry_delay=0.001,
-          retry_backoff=True, retry_kwargs=dict(max_retries=None),
-          shared=False)
-def send(self, payload):
-    """Send a VOEvent to GCN."""
-    payload = payload.encode('utf-8')
-    nbytes = len(payload)
-
-    conn = self.conn
-    self.conn = None
-
-    if conn is None:
-        log.info('creating new socket')
-        sock = socket.socket(socket.AF_INET)
-        try:
-            sock.bind((app.conf['gcn_bind_address'],
-                       app.conf['gcn_bind_port']))
-            sock.listen(0)
-            while True:
+    This is a single-client VOEvent broker (e.g. server). It listens for a
+    connection from address :obj:`~gwcelery.celery.Base.gcn_bind_address` and
+    port :obj:`~gwcelery.celery.Base.gcn_bind_port`.
+    """
+    with contextlib.closing(socket.socket(socket.AF_INET)) as sock:
+        sock.settimeout(1.0)
+        sock.bind((app.conf['gcn_bind_address'], app.conf['gcn_bind_port']))
+        sock.listen(0)
+        while not self.is_aborted():
+            try:
                 conn, (addr, _) = sock.accept()
-                if addr == app.conf['gcn_remote_address']:
-                    break
-                else:
-                    log.error('connection denied to remote host %s', addr)
-        finally:
-            sock.close()
+            except socket.timeout:
+                continue
+            if addr == app.conf['gcn_remote_address']:
+                log.info('accepted connection from remote host %s', addr)
+                break
+            else:
+                log.error('denied connection from remote host %s', addr)
+                conn.close()
+        else:  # self.is_aborted()
+            return
+
+    with contextlib.closing(conn):
         conn.setsockopt(socket.SOL_SOCKET, socket.SO_LINGER,
                         struct.pack('ii', 1, 0))
         conn.setsockopt(socket.SOL_SOCKET, socket.SO_SNDBUF, 1)
 
-    log.info('sending payload of %d bytes', nbytes)
-    try:
-        conn.sendall(_size_struct.pack(nbytes) + payload)
-    except:  # noqa
-        try:
-            conn.shutdown(socket.SHUT_RDWR)
-        except:  # noqa
-            log.exception('failed to shut down socket')
-        conn.close()
-        raise
-    self.conn = conn
+        while not self.is_aborted():
+            # Get next payload from queue in first-in, first-out fashion
+            payload = self.backend.lindex(_queue_name, 0)
+            if payload is None:
+                time.sleep(1)
+                continue
+            nbytes = len(payload)
+
+            log.info('sending payload of %d bytes', nbytes)
+            conn.sendall(struct.pack('!I', nbytes) + payload)
+            self.backend.lpop(_queue_name)
+
+
+_queue_name = broker.name + '.voevent-queue'
+
+
+@app.task(bind=True, ignore_result=True, shared=False)
+def send(self, payload):
+    """Send a VOEvent to GCN.
+
+    Under the hood, this task just pushes the payload onto a Redis queue,
+    and :func:`~gwcelery.tasks.gcn.broker` sends it."""
+    self.backend.rpush(_queue_name, payload.encode('utf-8'))
 
 
 class _VOEventDispatchHandler(DispatchHandler):