+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.0.19 (unreleased)
+## 0.0.19 (2018-07-09)
 
 -   Rewrite the GCN broker so that it does not require a dedicated worker.